octo/requirements.txt
PyQt5>=5.15.0
numpy>=1.20.0
pyyaml>=5.4.0
colorama>=0.4.4
